### 2024-06-20 #6908 @editorialbot
[PRE REVIEW]: DiffeRT2d: A Differentiable Ray Tracing Python Framework for Radio Propagation
**Submitting author:** @jeertmans (<a href="http://orcid.org/0000-0002-5579-5360">Jérome Eertmans</a>)
**Repository:** https://github.com/jeertmans/DiffeRT2d
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.2
