### 2024-06-21 #6908 @idoby
@danielskatz Hey, this came up in my GitHub feed and seems cool. I'd like to be a reviewer if possible.
