### 2024-06-21 #6908 @danielskatz
👋 @LyceanEM - Would you be willing to review this JOSS submission?
