### 2024-06-21 #6908 @danielskatz
👋 @roth-jakob, @hawkinsp, and @wsmoses - Given the JAX commonality here with the paper we just finished, I wonder if any of you would be interested in reviewing this submission?
