### 2024-06-23 #6915 @roth-jakob
Hi, thanks for your effort in writing the paper! Just a follow-up on the 4. above comment. I fully agree with that. Giving a little bit more explanation for readers not familiar with the subdiscipline could make the paper much more accessible.
