### 2024-06-23 jeertmans/DiffeRT2d#67 @roth-jakob
[DOC] Broken links in readme (JOSS Review)
Some links in the README are broken. When I click "Contributing guide" and "examples", I get a 404 Not Found message.

Ref: openjournals/joss-reviews#6915
