### 2024-06-25 b987b3b
fix(docs): broken links and add check
Closes #67
