### 2024-06-25 jeertmans/DiffeRT2d#67 @jeertmans
Hello @roth-jakob, thanks for catching those!

Keeping track of all links isn't always easy, even with automated tests (see #68).
I have fixed the broken links (I did a typo by adding `latest/` to the wrong link), but I am still thinking a bit how to fix other false positive errors reported by `markdown-link-check`.
