### 2024-06-26 cfeaacb
chore(paper): address comments from @idoby
This PR addresses the comments from @idoby, especially from https://github.com/openjournals/joss-reviews/issues/6915#issuecomment-2185022268.

Ref: https://github.com/openjournals/joss-reviews/issues/6915
