### 2024-06-26 #6915 @jeertmans
> Hey @jeertmans, thanks for submitting your package to JOSS! 📦 It seems very cool!

Hello @idoby, thanks for your first review comments, which I try to address in https://github.com/jeertmans/DiffeRT2d/pull/69.

Let me add a few comments here:

> A few comments about the paper:
>
> 1. Lines 50-51: "such as the number of multipath components **and** angle of arrival."
> 2. Line 68: Please spell out the acronym MPT for people who don't know what it means.
> 3. Line 69: "...we can easily accomodate **these surfaces**." Referring to objects here can be confusing since the next sentence refers to software objects rather than physical simulation objects.

Thanks for catching! I have rephrased those sentences.

> 4. Figure 2: "it is evident that..." It isn't evident. Please assume the reader is not well versed in the specific subdiscipline, explain in more detail what is being shown in the figure. Point out what the RIS is and where it is depicted in the image. I also advise to plot a specific path to illustrate clearly the 45-degree angle you're referring to. re: the code, I suggest to provide the code sample as a complete, runnable script with the package and to link externally to the relevant Python file. As it stands, the code is (admittedly) incomplete and not runnable because it omits the actual drawing of the figure. While understandable, this takes away substantial value from a potential user. Consider linking to a stable repository (e.g., Zenodo) as well as GitHub once the paper is accepted and your software archive is created. Besides this minor detail, I think the code illustrates the basic usage of the package well. As a final point, if you do decide to link to the complete example as an external file, consider pointing this out in the text and not as a footnote. If you do not want to include a link, it's enough to state that the example is available as a sample in the package repository.

I think I have addressed this issue by (1) putting the complete code and (2) rephrasing the caption.

> 5. Figure 3: Likewise, I suggest providing full runnable code for this example as well, though there is indeed no need to include the code in the paper itself.

Done.

> 6. "Machine Learning" section: It is unclear what "this COST meeting" means and why it is relevant to the paper. I suggest rewriting this part and including a representative figure from the linked page. I also strongly suggest linking to a stable URL and not your personal website.

I hope this is now clearer (I forgot to update that part before submitting, this is my bad).

I always hesitate between putting links from Zenodo, GitHub, or else. The motivation to use a link to my website was that, I am more likely to keep my `eertmans.be` up for a long time, than keeping the absolute path to the rendered notebook the same. On my website, I just set up a permalink redirect that I can update anytime I want. The permalink is also checked by tests, so hopefully I will always point to a valid url.
Zenodo (or web archives) can be nice to have a permalink, but then you cannot update the code (e.g., to fix typos etc.).

> Other than these minor comments, good work! I will take a look at the software itself soon.

Looking forward to the software review! :D
