### 2024-06-26 #6915 @jeertmans
> Hi, thanks for your effort in writing the paper! Just a follow-up on the 4. above comment. I fully agree with that. Giving a little bit more explanation for readers not familiar with the subdiscipline could make the paper much more accessible.

Hello! I have (I think) addressed this comment in https://github.com/jeertmans/DiffeRT2d/pull/69. Can you review it and tell me if it needs more details?
