### 2024-06-26 #6915 @danielskatz
Personally, I'm ok with content on website being linked to, but I would prefer that an archive.org copy of it was linked, as this is more "permanent" if possible.   e.g., rather than references https://www.ncsa.illinois.edu, one could reference https://web.archive.org/web/20240620232147/https://www.ncsa.illinois.edu/ for the version as of a few days ago.

For the link above, using the hdl (http://hdl.handle.net/2078/288635) would be better than using the direct link, even though it points to the same thing. If https://dial.uclouvain.be is an institutional repository, which I think it is, then this is enough. If not, then doing what @idoby suggests would be good, putting it on a "permanent" archive like Zenodo, arXiv, ...
