### 2024-06-26 610c8ba
feat(lib): add basic vertex diffraction
