+++ tests/test_geometry.py
@@ -154,6 +154,36 @@ class TestPoint:
         chex.assert_shape(got, (2, 2))
 
 
+class TestVertex:
+    @point
+    def test_parametric_to_cartesian(self, point: list):
+        vertex = Vertex(xy=jnp.array(point))
+        chex.assert_trees_all_equal(
+            vertex.parametric_to_cartesian(jnp.empty(0)), vertex.xy
+        )
+
+    @point
+    def test_cartesian_to_parametric(self, point: list):
+        vertex = Vertex(xy=jnp.array(point))
+        assert vertex.cartesian_to_parametric(jnp.empty(2)).size == 0
+
+    @point
+    @approx
+    def test_contains_parametric(self, point: list, approx: bool):
+        vertex = Vertex(xy=jnp.array(point))
+        assert is_true(
+            vertex.contains_parametric(jnp.empty(0), approx=approx), approx=approx
+        )
+
+    @point
+    @approx
+    def test_intersects_cartesian(self, point: list, approx: bool):
+        vertex = Vertex(xy=jnp.array(point))
+        assert is_false(
+            vertex.intersects_cartesian(jnp.empty((2, 2)), approx=approx), approx=approx
+        )
+
+
 class TestRay:
     @origin_dest
     def test_origin(self, origin: list, dest: list):
