### 2024-06-26 90af014
chore(paper): use bib reference instead
