### 2024-06-26 3f0fc06
chore(tests): added missed line coverage
