### 2024-06-26 #6915 @jeertmans
Indeed, https://dial.uclouvain.be/ is an institutional repository, and should be a permalink that will always be valid. This is a preliminary work, so not really published, but I transformed the URL-reference to a citation that points to the "preprint", available on our inst. website.

For the other internet links, I will convert them later (today or this week) to web archives :-)
