### 2024-06-27 81e7ffb
fix(docs): examples
