### 2024-06-27 #6915 @roth-jakob
Thanks for fixing the links and adding the full code for the figure in jeertmans/DiffeRT2d#69. I was able to reproduce the figure.  I will have a look at the software in the next few days.
