### 2024-06-27 #6915 @jeertmans
> Looks like the new paper proof did not pick up the changes

Yes, I forgot to first merge the PR ^^'. I'll fix that soon!
