### 2024-06-27 2cf5e77
chore(paper): address comments from @idoby (#69)
* chore(paper): address comments from @idoby

This PR addresses the comments from @idoby, especially from https://github.com/openjournals/joss-reviews/issues/6915#issuecomment-2185022268.

Ref: https://github.com/openjournals/joss-reviews/issues/6915

* chore(paper): use bib reference instead
