+++ papers/joss/paper.bib
@@ -73,6 +73,15 @@
   journal = {Differentiable Programming workshop at Neural Information Processing Systems 2021},
 }
 
+@misc{mlhelsinki,
+  author  = {Eertmans, Jérome and Oestges, Claude and Jacques, Laurent and others},
+  title   = {Learning to Sample Ray Paths for Faster Point-to-Point Ray Tracing},
+  year    = {2024},
+  month   = {jun},
+  journal = {COST INTERACT 8th Meeting (Helsinki, from 2024/06/17 to 2024/06/20)},
+  url     = {http:// hdl.handle.net/2078/288635},
+}
+
 @inproceedings{mpt-eucap2023,
   author    = {Eertmans, J\'{e}rome and Oestges, Claude and Jacques, Laurent},
   title     = {{M}in-{P}ath-{T}racing: {A} {D}iffraction {A}ware {A}lternative to {I}mage {M}ethod in {R}ay {T}racing},
