### 2024-06-27 7aa9b3b
Typos and phrasing in jax_and_jaxtyping.md
