### 2024-06-27 53a83bb
fix(docs): broken links and add check (#68)
* fix(docs): broken links and add check

Closes #67

* chore(ci): fixing broken links
