+++ pyproject.toml
@@ -86,7 +86,7 @@ filename = "CHANGELOG.md"
 replace = '''<!-- start changelog -->
 
 (unreleased)=
-## [Unreleased](https://github.com/jeertmans/DiffeRT2d/compare/{new_version}...HEAD)'''
+## [Unreleased](https://github.com/jeertmans/DiffeRT2d/compare/v{new_version}...HEAD)'''
 search = "<!-- start changelog -->"
 
 [tool.codespell]
