### 2024-06-27 0af5f2b
chore(lib): changing default color
