### 2024-06-27 e1dc798
fix(docs): typos and phrasing in `docs/source/contributing/index.md` (#74)
