### 2024-06-27 f6ddb3c
fix(docs): typos and phrasing in `docs/source/quickstart.md` (#73)
