### 2024-06-27 59822cd
fix(docs): typos and phrasing in `docs/source/index.md` (#71)
