### 2024-06-27 88cca06
Update jax_and_jaxtyping.md
