### 2024-06-27 ab6f9fb
fix(docs): typos and phrasing in `docs/source/jax_and_jaxtyping.md` (#72)
* Typos and phrasing in jax_and_jaxtyping.md

* Update jax_and_jaxtyping.md
