### 2024-06-27 #6915 @jeertmans
> The software itself looks good to me, pretty cool work. I'm not a huge fan of a fully object-oriented style with abstract classes but that's really a matter of preference/design choices. I have no comments on the software other than good job, it seems well crafted and easy to read. 👍

Thanks!

> I did find some typos in the documentation, I will open a PR to fix them soon.

I have reviewed and merge all PRs, thanks for your help!

> Other than that, l think the doc intro page deserves a little more exposition wrt what this tool is and how it's different than other similar tools (what JOSS would call a "statement of need"). I suggest to take some prose from the paper and graduate the footnote on the intro page into normal text. However, I leave this up to the author and will not block the review over this.

Good remark, I will wait for the second reviewer (and the editor?) to give their feedback first, then I think I might improve a bit the documentation according to your suggestions.

Other than that, I will re-generate a paper with the latest changes.
