### 2024-06-27 #6915 @jeertmans
My TODO list:

- [ ] Improve the main `index.md` documentation page, as suggested by @idoby
- [ ] Create a new Zenodo *release* to point to the latest version (and update in `README.md`)
- [ ] Change URLs in paper to webarchive links (or Zenodo)
- [ ] Change package version in JOSS paper (to include latest changes)

> I put that here to keep track of important changes to make, and I won't have access to a computer from July 2nd to July 16th, so this doesn't get lost :-)
