### 2024-06-27 3c54745
fix(examples): set_offsets
