### 2024-06-27 59cdf8f
feat(lib): add basic vertex diffraction (#70)
* feat(lib): add basic vertex diffraction

* fix(tests): correct

* chore(tests): improve coverage

* chore(tests): added missed line coverage

* fix(docs): examples

* fix(docs): typo

* fix(docs): add missing key argument

* chore(lib): changing default color

* chore(docs): re-order changelog

* fix(examples): set_offsets
