+++ differt2d/scene.py
@@ -4,12 +4,11 @@ __all__ = ("Scene", "SceneName")
 
 import json
 from abc import abstractmethod
-from collections.abc import Iterator, Mapping, Sequence
+from collections.abc import Callable, Iterator, Mapping, Sequence
 from functools import singledispatchmethod
 from itertools import groupby, product
 from typing import (
     Any,
-    Callable,
     Generic,
     Literal,
     Optional,
@@ -23,7 +22,7 @@ import equinox as eqx
 import jax
 import jax.numpy as jnp
 from beartype import beartype as typechecker
-from differt_core.rt.graph import CompleteGraph
+from differt_core.rt.graph import CompleteGraph, DiGraph
 from jaxtyping import Array, Float, Int, PRNGKeyArray, jaxtyped
 from matplotlib.artist import Artist
 
@@ -100,6 +99,59 @@ class PyTreeDict(eqx.Module, Mapping[_K, _V]):
         return len(self._keys)
 
 
+@eqx.filter_jit
+def all_path_candidates(
+    num_nodes: int,
+    min_order: int = 0,
+    max_order: int = 1,
+    *,
+    order: Optional[int] = None,
+    filter_nodes: Optional[tuple[int, ...]] = None,
+) -> list[Int[Array, "num_path_candidates order"]]:
+    """
+    Returns all path candidates, as a list of array of indices.
+
+    This function is a lower-level version of :meth:`Scene.all_path_candidates`
+    that utilizes caching for more efficient computations.
+
+    In practice, you should not use this function directly, but rather
+    :meth:`Scene.all_path_candidates`.
+
+    :param num_nodes: The number of nodes, i.e., the objects,
+        present in the `graph`. This excludes the transmitting
+        and receiving nodes.
+    :param min_order: The minimum order of the path, i.e., the
+        number of interactions.
+    :param max_order: The maximum order of the path, i.e., the
+        number of interactions.
+    :param order: If provided, it is equivalent to setting
+        ``min_order=order`` and ``max_order=order``.
+    :param filter_nodes: An optional list of nodes, i.e., object indices,
+        to not visit.
+    :return: The list of list of indices.
+    """
+    if filter_nodes is None:
+        graph = CompleteGraph(num_nodes)
+        from_ = num_nodes
+        to = from_ + 1
+    else:
+        graph = DiGraph.from_complete_graph(CompleteGraph(num_nodes))
+        from_, to = graph.insert_from_and_to_nodes()
+        graph.disconnect_nodes(*filter_nodes)
+
+    if order is not None:
+        min_order = order
+        max_order = order
+
+    return [
+        jnp.asarray(path_candidate, dtype=jnp.int32)
+        for order in range(min_order, max_order + 1)
+        for path_candidate in graph.all_paths(
+            from_, to, order + 2, include_from_and_to=False
+        )
+    ]
+
+
 class Scene(Plottable, eqx.Module, Generic[_O]):
     """2D Scene made of objects, one or more transmitting node(s), and one or more receiving node(s)."""
 
@@ -149,6 +201,58 @@ class Scene(Plottable, eqx.Module, Generic[_O]):
         :return: The new scene.
         """
         return eqx.tree_at(lambda s: s.objects, self, tuple(objects))
+
+    @jaxtyped(typechecker=typechecker)
+    def filter_objects(self, filter_spec: Callable[[Object], bool]) -> "Scene":
+        """
+        Returns a copy of this scene, with the given objects filtered out.
+
+        :param filter_spec: A callable indicating
+            which objects should be kept.
+        :return: The new scene.
+
+        :Examples:
+
+        A practical use case for this function is to plot a basic scene,
+        but only perform ray tracing on a subset of the objects, e.g.,
[…]
