+++ .github/workflows/test.yml
@@ -33,7 +33,13 @@ jobs:
       run: rye sync
 
     - name: Run Pytest
-      run: rye run pytest -n auto
+      run: rye run pytest
+
+    - name: Upload to codecov.io
+      uses: codecov/codecov-action@v4
+      with:
+        token: ${{ secrets.CODECOV_TOKEN }}
+        fail_ci_if_error: true
 
   benchmark:
     runs-on: ubuntu-latest
