+++ .github/workflows/coverage.yml
@@ -1,37 +1,0 @@
-on: [push]
-
-name: Code Coverage
-
-jobs:
-  test:
-    name: Coverage
-    runs-on: ubuntu-latest
-    steps:
-    - name: Checkout repository
-      uses: actions/checkout@v4
-
-    - name: Setup Rye
-      if: matrix.os != 'windows-latest'
-      env:
-        RYE_TOOLCHAIN_VERSION: '3.11'
-        RYE_INSTALL_OPTION: --yes
-      run: |
-        curl -sSf https://rye.astral.sh/get | bash
-        echo "$HOME/.rye/shims" >> $GITHUB_PATH
-
-    - name: Configure Rye
-      run: |
-        rye config --set-bool behavior.use-uv=true
-        rye pin '3.11'
-
-    - name: Install Differt2D
-      run: rye sync
-
-    - name: Run pytest and coverage
-      run: rye run pytest --cov-report xml --cov=differt2d tests/ differt2d/
-
-    - name: Upload to codecov.io
-      uses: codecov/codecov-action@v4
-      with:
-        token: ${{ secrets.CODECOV_TOKEN }}
-        fail_ci_if_error: true
