### 2024-06-28 #6915 @roth-jakob
@jeertmans, nice software package! I had a closer look at it. I have only one minor suggestion, along the lines of @idoby's comment. I personally like it if there are a few sentences at the top of the README explaining what the package is doing and for whom it is targeted. All this information is in the paper, and you could consider summarizing it at the top of the README.
