### 2024-06-28 #6915 @danielskatz
👋 @jeertmans

> My TODO list:
>
> * [ ]  Improve the main `index.md` documentation page, as suggested by @idoby
> * [ ]  Create a new Zenodo _release_ to point to the latest version (and update in `README.md`)
> * [ ]  Change URLs in paper to webarchive links (or Zenodo)
> * [ ]  Change package version in JOSS paper (to include latest changes)
>
> > I put that here to keep track of important changes to make, and I won't have access to a computer from July 2nd to July 16th, so this doesn't get lost :-)

At the end of the process, once the review is complete, I will proofread the paper, and perhaps suggest changes, then ask you to make an archival deposit (e.g., in Zenodo) and tell me that DOI and the version associated with it, so you might want to wait on the 2nd and 4th items above until then.
