### 2024-06-28 #6915 @roth-jakob
Yes, for me, it's the same. I proposed to @jeertmans to expand the readme a bit, but otherwise, I'm done with the review.
