### 2024-06-28 2121e98
chore(paper): permanent links and README intro
