+++ docs/source/index.md
@@ -6,31 +6,9 @@ og:description: 2D Toolbox for Differentiable Ray Tracing
 # Welcome to DiffeRT2d's documentation
 
 DiffeRT2d is a Python toolbox for 2D differentiable Ray Tracing,
-with a focus on Radio Propagation applications[^1].
-
-[^1]: This means we are mostly interested in simulating paths from
-  one node (transmitter) to another (receiver),
-  i.e., Point-to-Point Ray Tracing.
-  DiffeRT2d is not designed as a highly efficient image renderer, like in computer graphics.
-  A renderer will usually favor a Ray Launching method instead.
-
-The present tool is thoroughly documented, so please have a look at the
-following sections:
-
-```{toctree}
-:titlesonly:
-:maxdepth: 1
-
-quickstart
-examples_gallery/index
-reference/index
-research/index
-jax_and_jaxtyping
-references
-```
-
-If you are interested in contributing to this tool, please checkout the
-[Contributing](contributing/index) section!
+with a focus on Radio Propagation applications, where we are mostly
+interested in simulating paths from one node (transmitter)
+to another (receiver), i.e., Point-to-Point Ray Tracing.
 
 ```{eval-rst}
 .. plot::
@@ -58,6 +36,30 @@ If you are interested in contributing to this tool, please checkout the
     scene.plot(ax)
     plt.show()  # doctest: +SKIP
 ```
+
+
+```{include} ../../README.md
+:start-after: <!-- start description -->
+:end-before: <!-- end description -->
+```
+
+The present tool is thoroughly documented, so please have a look at the
+following sections:
+
+```{toctree}
+:titlesonly:
+:maxdepth: 1
+
+quickstart
+examples_gallery/index
+reference/index
+research/index
+jax_and_jaxtyping
+references
+```
+
+If you are interested in contributing to this tool, please checkout the
+[Contributing](contributing/index) section!
 
 ```{toctree}
 :caption: Development
