### 2024-06-28 8d3561d
chore(paper): permanent links and README intro (#77)
* chore(paper): permanent links and README intro

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
