+++ differt2d/__version__.py
@@ -1,3 +1,3 @@
 """Module version."""
 
-__version__ = "0.3.2"
+__version__ = "0.3.3"
