+++ pyproject.toml
@@ -44,7 +44,7 @@ Repository = "https://github.com/jeertmans/DiffeRT2d"
 allow_dirty = false
 commit = true
 commit_args = ""
-current_version = "0.3.2"
+current_version = "0.3.3"
 ignore_missing_version = false
 message = "chore(deps): bump version from {current_version} to {new_version}"
 parse = '(?P<major>\d+)\.(?P<minor>\d+)\.(?P<patch>\d+)(-rc(?P<release>\d+))?'
