+++ CITATION.cff
@@ -22,7 +22,7 @@ keywords:
   - JAX
   - Radio Propagation
 license: MIT
-version: v0.3.2
+version: v0.3.3
 preferred-citation:
   publisher:
     name: IEEE
