### 2024-06-28 #6915 @jeertmans
@danielskatz done, the latest version is v0.3.3 (https://github.com/jeertmans/DiffeRT2d/releases/tag/v0.3.3). I'll update the Zenodo after you have proofread the paper, as you suggested :-)
