### 2024-06-28 #6915 @danielskatz
My institution's summer picnic is starting soon, but I'll do this after it and some meetings, in the 4-8 hours...
