### 2024-06-28 97b99e5
minor changes in paper.bib
