+++ papers/joss/paper.bib
@@ -84,7 +84,7 @@
 
 @inproceedings{mpt-eucap2023,
   author    = {Eertmans, J\'{e}rome and Oestges, Claude and Jacques, Laurent},
-  title     = {{M}in-{P}ath-{T}racing: {A} {D}iffraction {A}ware {A}lternative to {I}mage {M}ethod in {R}ay {T}racing},
+  title     = {{M}in-{P}ath-{T}racing: A Diffraction Aware Alternative to Image Method in Ray Tracing},
   year      = {2023},
   booktitle = {2023 17th European Conference on Antennas and Propagation (EuCAP)},
   volume    = {},
@@ -110,7 +110,7 @@
 
 @software{pylayers,
   author  = {Uguen, Bernard and Amiot, Nicolas and Laaraiedh, Mohamed and Mhedhbi, Meriem and Avrillon, Stephane and Burghelea, Roxana and Plouhinec, Eric and Talom, Friedman Tchoffo and Chaluyman, Taguhi and Lei, Yu},
-  title   = {{A}dvanced {R}adio {C}hannel {S}imulator},
+  title   = {Advanced Radio Channel Simulator},
   year    = {2014},
   url     = {http://github.com/pylayers/pylayers},
   version = {0.5},
