### 2024-06-28 #6915 @danielskatz
@jeertmans - I've suggested some changes in https://github.com/jeertmans/DiffeRT2d/pull/78. Please merge this or let me know what you disagree with.
