### 2024-06-29 f49f239
chore(paper): JOSS changes in paper and bib (#78)
* minor changes in paper.md

* minor changes in paper.bib
