### 2024-06-29 7a56dca
chore(paper): removed useless footnote
