### 2024-06-29 #6915 @jeertmans
Thanks, @danielskatz! It looked great, and I also remove one footnote that was no longer needed (as I now include the full code snippet).
