### 2024-06-30 ecbe7b9
fixing a bug in paper.md
