### 2024-06-30 #6915 @danielskatz
@jeertmans - I seem to have introduced a bug - sorry. https://github.com/jeertmans/DiffeRT2d/pull/79 should fix it. Once you merge this, please check the proof to make it's ok, and once it is, please:
