+++ CHANGELOG.md
@@ -8,7 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- start changelog -->
 
 (unreleased)=
-## [Unreleased](https://github.com/jeertmans/DiffeRT2d/compare/v0.3.3...HEAD)
+## [Unreleased](https://github.com/jeertmans/DiffeRT2d/compare/v0.3.4...HEAD)
+
+(v0.3.4)=
+## [v0.3.4](https://github.com/jeertmans/DiffeRT2d/compare/v0.3.3...v0.3.4)
 
 (v0.3.3)=
 ## [v0.3.3](https://github.com/jeertmans/DiffeRT2d/compare/v0.3.2...v0.3.3)
