### 2024-06-30 #6915 @jeertmans
There you go @danielskatz: [10.5281/zenodo.12600658](https://doi.org/10.5281/zenodo.12600658).

I am just not sure: should the Zenodo metadata also contain the paper title and authors?
