### 2024-06-30 #6915 @jeertmans
> The title and authors of the Zenodo deposit should be the same as those of the paper. You can change these manually without making a new deposit by editing the metadata.

Done!
