### 2024-06-30 #6915 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
