### 2024-06-30 #6915 @danielskatz
@jeertmans - it looks like there's a valid DOI missing from your bib file, and in fact, you're citing a preprint rather than the later published article.  Can you change the entry for sionnart to this?
```
@INPROCEEDINGS{sionnart,
  author={Hoydis, Jakob and Aoudia, Faycal Ait and Cammerer, Sebastian and Nimier-David, Merlin and Binder, Nikolaus and Marcus, Guillermo and Keller, Alexander},
  booktitle={2023 IEEE Globecom Workshops ({GC W}kshps)},
  title={Sionna {RT}: Differentiable Ray Tracing for Radio Propagation Modeling},
  year={2023},
  pages={317--321},
  doi={10.1109/GCWkshps58843.2023.10465179},
}
```

You don't need to update the zenodo deposit, as we don't really require the paper source to be there at all - we archive the pdf and XML versions that we create.
