### 2024-06-30 #6915 @danielskatz
@jeertmans - also, in the caption for Figure 3, please change
```see [@eertmans2024eucap] for all details``` to ```see @eertmans2024eucap for all details```
