+++ papers/joss/paper.bib
@@ -116,13 +116,16 @@
   version = {0.5},
 }
 
-@misc{sionnart,
-  author        = {Hoydis, Jakob and Aoudia, Fay\c{c}al A\"{\i}t and Cammerer, Sebastian and Nimier-David, Merlin and Binder, Nikolaus and Marcus, Guillermo and Keller, Alexander},
-  title         = {Sionna RT: Differentiable Ray Tracing for Radio Propagation Modeling},
-  year          = {2023},
-  eprint        = {2303.11103},
-  archiveprefix = {arXiv},
-  primaryclass  = {cs.IT},
+@inproceedings{sionnart,
+  author    = {Hoydis, Jakob and Aoudia, Faycal Ait and Cammerer, Sebastian and Nimier-David, Merlin and Binder, Nikolaus and Marcus, Guillermo and Keller, Alexander},
+  booktitle = {2023 IEEE Globecom Workshops (GC Wkshps)},
+  title     = {Sionna RT: Differentiable Ray Tracing for Radio Propagation Modeling},
+  year      = {2023},
+  volume    = {},
+  number    = {},
+  pages     = {317--321},
+  keywords  = {Computational modeling;Propagation;Radio transmitters;Transmitting antennas;Receiving antennas;Ray tracing;Reconfigurable intelligent surfaces},
+  doi       = {10.1109/GCWkshps58843.2023.10465179},
 }
 
 @misc{tensorflow,
