### 2024-06-30 #6915 @danielskatz
Congratulations to @jeertmans ([Jérome Eertmans](http://orcid.org/0000-0002-5579-5360)) and co-authors on your publication!!

And thanks to @idoby and @roth-jakob for a quick and successful review process!
JOSS depends on volunteers, and wouldn't be possible without efforts like yours
